Implementada parte de Seguranca e Verificação por Token

+++ src/main/java/com/inatel/projeto/controller/GameController.java
@@ -5,11 +5,17 @@ package com.inatel.projeto.controller;
 
 
 import java.net.URI;
-import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort.Direction;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -18,6 +24,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.util.UriComponentsBuilder;
 
@@ -25,7 +32,7 @@ import com.inatel.projeto.controller.dto.GameDto;
 import com.inatel.projeto.controller.form.GameForm;
 import com.inatel.projeto.model.Game;
 import com.inatel.projeto.repository.GameRepository;
-import com.inatel.projeto.repository.PromocaoRepository;
+
 
 @RestController
 @RequestMapping(path="/games")
@@ -34,11 +41,12 @@ public class GameController {
 	@Autowired
 	private GameRepository gamerepository;
 	
-	@Autowired
-	private PromocaoRepository promocaorepository;
+	
 	
 	
 	@PostMapping
+	@Transactional
+	@CacheEvict(value="listaDeJogos",allEntries = true)
 	public  ResponseEntity<GameDto> adicionarNovoGame(@RequestBody GameForm form,  UriComponentsBuilder uriBuilder) {
 		Game game = form.converter();
 		gamerepository.save(game);
@@ -49,12 +57,14 @@ public class GameController {
 		
 		}
 	@GetMapping
-	public List<GameDto> lista(String name){
+	@Cacheable(value="listaDeJogos")
+	public Page<GameDto> lista(@RequestParam(required = false) String name, @PageableDefault(sort = "Name",direction = Direction.ASC, page = 0, size = 10) Pageable paginacao){
+		
 		if (name == null) {
-			List<Game> games =  gamerepository.findAll();
+			Page<Game> games =  gamerepository.findAll(paginacao);
 			return GameDto.converter(games);
 		}else {
-			List<Game> games =  gamerepository.findByName(name);
+			Page<Game> games =  gamerepository.findByName(name,paginacao);
 			return GameDto.converter(games);
 			
 		}
@@ -64,28 +74,56 @@ public class GameController {
 	
 	
 	@GetMapping("/{id}")
-    public GameDto detalhar( @PathVariable Integer id) {
+    public ResponseEntity<GameDto> detalhar( @PathVariable Integer id) {
 	
-		Game game = gamerepository.getOne(id);
+		Optional<Game> game = gamerepository.findById(id);
 	
+		if (game.isPresent()) {
+			return ResponseEntity.ok( new GameDto(game.get()));
+		}else {
+			
+			return ResponseEntity.notFound().build();
+		}
 		
-		return new GameDto(game);
+		
 
 
 }
 	@PutMapping("/{id}")
 	@Transactional
+	@CacheEvict(value="listaDeJogos",allEntries = true)
 	public ResponseEntity<GameDto> atualizar(@PathVariable Integer id,@RequestBody GameForm form){
-		Game game = form.atualizar(id,gamerepository);
+		
+		Optional<Game> optional = gamerepository.findById(id);
+		
+		if (optional.isPresent()) {
+			Game game = form.atualizar(id,gamerepository);
 		return ResponseEntity.ok(new GameDto(game));
+			
+		}else {
+			
+			return ResponseEntity.notFound().build();
+		}
+		
 		
 	}
 	
 	@DeleteMapping("/{id}")
 	@Transactional
+	@CacheEvict(value="listaDeJogos",allEntries = true)
 	public ResponseEntity<?> remover(@PathVariable Integer id){
-		gamerepository.deleteById(id);
-		return ResponseEntity.ok().build();
+         
+		Optional<Game> optional = gamerepository.findById(id);
+		
+		if (optional.isPresent()) {
+		    gamerepository.deleteById(id);
+		    return ResponseEntity.ok().build();
+			
+		}else {
+			
+			return ResponseEntity.notFound().build();
+		}
+		
 	}
